move the hyperv-only powershell dependency into the hyperv subdirectory
