+++ builder/hyperv/common/driver_ps_4.go
@@ -8,8 +8,8 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/hashicorp/packer/common/powershell"
-	"github.com/hashicorp/packer/common/powershell/hyperv"
+	"github.com/hashicorp/packer/builder/hyperv/common/powershell"
+	"github.com/hashicorp/packer/builder/hyperv/common/powershell/hyperv"
 )
 
 type HypervPS4Driver struct {
