+++ builder/hyperv/common/powershell/powershell.go
