+++ builder/hyperv/common/powershell/powershell_test.go
