+++ builder/hyperv/common/powershell/scriptbuilder.go
